+++ CHANGES.md
@@ -5,3 +5,4 @@
 ## 0.2.4
 * changing deprecated `std::env::home_dir` to `dirs::home_dir`
 * improved the efficiency of --strict and --prove
+* added optimization step so anagram discovery uses a trie containing only words embeddable in the given phrase